I feel sad about this, but... :)
Here is a short story...
I used sublime for a long time, than I wanted to try out vs code...
There was a missing extension from sublime that I really liked and that was PHP companion...
It was the end of 2017...
I wanted to build a plugin like that, but I didn't know how...
So I started to look at the vs code API... 
It took me days to read it...
Than I just started and there it was... My first prototype!
It worked OK, but it wasn't meant for other people, it had a lot of bugs...
So I tired to make it better and tried, but every time a new problem would face up...
I would always get stuck at some point...
30 January 2018, I gave up...
That night I went to my friend, she was at work...
She told me that she had a dream, that I made something really cool...
And that I became successful :D
I told her that I was building something... but I couldn't make it the way I wanted to
So I came home...
And the next two days, I tried again and again...
And to my surprise, I turn out to be exactly how I imagined it :)

I published it 2/4/2018
That day I noticed a guy publish a similar extension called `php constructor` :D 
He was noticed by the Laravel community, and Laracast made a tweet about it :D 
At that moment, I tough, that I could be that guy if I just made it a couple of days earlier :D but oh well

Because I didin't wanted to brag that I build a cool extension :sunglasses: I relied on quality
So I implemented some other cool features, like method generation, getters and setters...
And I was happy with it...
Other people noticed this plugin as well, I checked the download status every hour :D 
I am really glad I did this. It was a fun project I did in my free time.

But why did I stop using VS Code? 
VS code and Chrome ate my RAM (4GB), my laptop would just freeze. So I went back to Sublime.

Because I use Sublime, I don't feel the joy to maintain something that I don't use.
Sorry but that is the way things are :) 

Hope the story was interesting:)